requirements.txt
google-cloud-storage==1.28.1
Pillow==7.1.2
requests==2.23.0
